Add REPROJ_FIX_DEPTH matching guess type
This assumes all newly detected left camera features are at a fixed depth.
Then, those 3D features get reprojected to the right camera and the pixel where
they land is the initial guess for the optical flow optimizer.
This heuristic alone enables cameras which are not very parallel work,
as in these the SAME_PIXEL guess never converges. An example of datasets like
these can be found in the Odyssey+ datasets in bit.ly/monado-datasets.