Use IMU for feature tracking initial estimates
This commit introduces two new input queues in the frame-to-frame optical
flow module. One for receiving the latest estimated poses from the optimization
module. And the other for receiving IMU measurements.

Feature tracking now uses the latest state and preintegrates IMU samples to
estimate where a feature from the previous frame will be in the current one.

Before this, we used as the initial estimate the pixel coordinates of the feature
in the previous frames. For aggressive movements, the pixel distance between the
previous observation and the current one were usually too large for optical flow
to converge and the feature track was lost.